+++ test/pipelining_commands_test.rb
@@ -198,6 +198,19 @@ class TestPipeliningCommands < Minitest::Test
     assert_equal result.first, { "field" => "value" }
   end
 
+  def test_zpopmax_in_a_pipeline_produces_future
+    target_version('5.0.0') do
+      r.zadd("sortedset", 1.0, "value")
+      future = nil
+      result = r.pipelined do
+        future = r.zpopmax("sortedset")
+      end
+
+      assert_equal [["value", 1.0]], result
+      assert_equal ["value", 1.0], future.value
+    end
+  end
+
   def test_keys_in_a_pipeline
     r.set("key", "value")
     result = r.pipelined do |p|